requirements/base.txt
# Install with: `pip install -r requirements.txt`

Django >= 1.5.2
bpython>=0.12
django-braces==1.2.2
django-model-utils==1.4.0
South==0.8.1
Pillow==2.1.0
#MySQL-python==1.2.4
python-memcached >= 1.53, < 2.0

#CMS
django-cms >= 2.4.2
django-filer>=0.9.5

# Support libs
django-crispy-forms >= 1.3.2, < 1.4
django_compressor >= 1.3, < 1.4
-e git+https://github.com/clintecker/django-google-analytics.git@140787ba86ef4995b2cabf927bf0b56ec7728348#egg=google_analytics-0.2

django-vkontakte-iframe>=0.4.1
django-annoying==0.7.7
djangorestframework==2.3.7
